@@ -1,36 +1,45 @@
 {
-   "name": "inferno-isomorphic-tempalte",
-   "scripts": {
-      "dev": "set NODE_ENV=development && node -r ts-node/register --inspect fuse.ts dev",
-      "test": "node -r ts-node/register --inspect fuse.ts test"
-   },
-   "dependencies": {
-      "cookie-parser": "^1.4.3",
-      "express": "~4.16.0",
-      "inferno": "^5.0.6",
-      "inferno-create-element": "^5.0.6",
-      "inferno-router": "^5.0.6",
-      "inferno-server": "^5.0.6"
-   },
-   "devDependencies": {
-      "@types/cookie-parser": "^1.4.1",
-      "@types/enzyme": "^3.1.10",
-      "@types/express": "^4.11.1",
-      "@types/jest": "^23.0.0",
-      "@types/node": "^10.3.0",
-      "enzyme": "^3.3.0",
-      "enzyme-adapter-inferno": "^1.3.0",
-      "fuse-box": "^3.2.2",
-      "fuse-test-runner": "^1.0.16",
-      "inferno-devtools": "^5.0.6",
-      "inferno-test-utils": "^5.0.6",
-      "jest": "^23.1.0",
-      "jsdom": "11.11.0",
-      "jsdom-global": "3.0.2",
-      "ts-node": "^6.1.0",
-      "ts-transform-classcat": "^0.1.0",
-      "ts-transform-inferno": "^2.2.0",
-      "tslint-react-recommended": "^1.0.15",
-      "typescript": "^2.9.1"
-   }
+  "name": "inferno-isomorphic-template",
+  "license": "MIT",
+  "scripts": {
+    "check": "tsc",
+    "build:server": "webpack --env.platform=server",
+    "build:client": "webpack --env.platform=client",
+    "build": "yarn run build:server & yarn run build:client",
+    "serve": "node dist/js/server.js",
+    "start": "yarn run build && yarn run serve",
+    "dev": "nodemon --watch ./src/ -e ts,tsx,css,scss --exec yarn run start"
+  },
+  "dependencies": {
+    "express": "~4.17.1",
+    "inferno": "^7.4.3",
+    "inferno-create-element": "^7.4.3",
+    "inferno-hydrate": "^7.4.3",
+    "inferno-router": "^7.4.3",
+    "inferno-server": "^7.4.3"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.5.5",
+    "@babel/plugin-transform-typescript": "^7.11.0",
+    "@babel/preset-env": "7.11.5",
+    "@babel/preset-typescript": "^7.3.3",
+    "@types/cookie-parser": "^1.4.1",
+    "@types/express": "^4.17.8",
+    "@types/jest": "^26.0.13",
+    "@types/node": "^14.6.4",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-inferno": "^6",
+    "clean-webpack-plugin": "^3.0.0",
+    "cookie-parser": "^1.4.3",
+    "css-loader": "^4.2.2",
+    "mini-css-extract-plugin": "^0.11.0",
+    "node-sass": "^4.12.0",
+    "nodemon": "^2.0.4",
+    "sass-loader": "^10.0.1",
+    "style-loader": "^1.2.1",
+    "typescript": "^4.0.2",
+    "webpack": "4.44.1",
+    "webpack-cli": "^3.3.6",
+    "webpack-dev-server": "3.11.0"
+  }
 }
